DeepHyperion-MNIST/requirements.txt
absl-py==0.9.0
bleach==1.5.0
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
cssselect2==0.3.0
cycler==0.10.0
deap==1.2.2
decorator==4.4.2
defusedxml==0.6.0
gobject==0.1.0
grpcio==1.16.1
h5py==2.9.0
html5lib==0.9999999
imageio==2.9.0
joblib==0.14.1
Keras==2.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.0.2
mock==4.0.2
networkx==2.4
numpy==1.18.3
opencv-python==4.3.0.36
pandas==1.0.3
Pillow==7.2.0
protobuf==3.11.3
pycairo==1.19.1
pycparser==2.19
Pygments==2.3.1
PyGObject==3.36.1
pyparsing==2.4.7
pypotrace==0.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sklearn==0.0
svgpathtools==1.3.3
svgwrite==1.4
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
threadpoolctl==2.1.0
tifffile==2020.7.4
tinycss2==1.0.2
tqdm==4.46.0
webencodings==0.5.1
Werkzeug==1.0.1
